requirements.txt
acres==0.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
filelock==3.18.0
httplib2==0.22.0
idna==3.10
isodate==0.6.1
looseversion==1.3.0
lxml==5.4.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
prov==2.0.1
puremagic==1.29
pydot==4.0.0
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
traits==7.0.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
